% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ukhd.R
\name{scale_fill_ukhd}
\alias{scale_fill_ukhd}
\alias{scale_colour_ukhd}
\title{Sequential and diverging colour scales from the UKHD colour palettes}
\usage{
scale_fill_ukhd(
  name = waiver(),
  ...,
  type = "seq",
  palette = "primary",
  direction = 1,
  na.value = "grey50"
)

scale_colour_ukhd(
  name = waiver(),
  ...,
  type = "seq",
  palette = 1,
  direction = 1,
  na.value = "grey50"
)
}
\arguments{
\item{type}{Type of palette. Options are "seq" (sequential) and "div"
(diverging)}

\item{palette}{Palette chosen. Options are "primary" (dark blue), "secondary"
(cyan), "grey", "orange", "beige", and "green" as well as "blueorange1",
"blueorange2", and "greige". Defaults to primary}

\item{direction}{Direction of the palette. Defaults to 1

For a continous version use the colours from ukhd_pal in scale_x_gradientn}
}
\description{
Sequential and diverging colour scales from the UKHD colour palettes
}
\author{
mariabnd
}
