% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ukhd.R
\name{theme_ukhd}
\alias{theme_ukhd}
\title{Add UKHD theme to ggplot chart}
\usage{
theme_ukhd(base_size = 16)
}
\arguments{
\item{base_size}{Desired base font size. Other font sizes are defined in
relation to it. Defaults to 16}
}
\description{
This function ensures adherence to the UKHD styleguide available from:
\href{https://design.ukhd.de}{Markenportal}
}
\author{
mariabnd
}
