% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ukhd.R
\name{ukhd_pal}
\alias{ukhd_pal}
\title{Colour palettes following the UKHD style guide}
\usage{
ukhd_pal(n, name = "primary")
}
\arguments{
\item{n}{Number of colours desired}

\item{name}{Palette chosen. Options are "primary" (dark blue), "secondary"
(cyan), "grey", "orange", "beige", and "green" as well as "blueorange1",
"blueorange2", and "greige". Defaults to primary}
}
\description{
This function provides colourschemes from the UKHD styleguide available from:
\href{https://design.ukhd.de}{Markenportal}
}
\examples{
scales::show_col(ukhd_pal())
colorblindr::palette_plot(ukhd_pal())
}
\author{
mariabnd
}
